Fix unique key list warning
Multiple ways to resolve this, but using an unique
id from the data is the most accepted route. Therefore instead of adding a library to create a uuid or similar, added a id to the data manually (mock the concept)

Important for items to have unique ids, because if your array items can move (e.g. due to sorting), get inserted, or get deleted. A well-chosen key helps React infer what exactly has happened, and make the correct updates to the DOM tree.