Removed grow transition for AvatarMenu.
Due to multiple issues that occur on the MUI transition in this instance, the grow transition is disabled.
This isn't the optimal solution, further investigation is needed.

Attempted to apply the solutions mentioned on bug reports. Added a React.Fragment wrap, which resolved only some of the issues. Specifically the "Uncaught TypeError: Cannot read properties of null (reading 'scrollTop').

Might be missing something small. See
https://github.com/mui/material-ui/issues/27154
https://github.com/mui/material-ui/issues/18119
https://github.com/mui/material-ui/issues/17119