Add html markdown to the home intro text.
The home intro text has a html bold tag which needs
to display. Added the MUI Typographys dangouslySetInnerHTML property. Which will then handle this display of the html tag.

Its possible to add other libraries to do this, but this route is the least invasive and doesn't require further libraries.

Its internally set text, so nothing "dangerous" that one would need to consider when its users adding html.