Add AccessDenied navigation and routing.
A mock sample was setup on the AvatarMenu to navigate a user to a AccessDenied routing when clicking on the Edit Profile option. This just showcases navigation setup using useHistory.

AccessDenied component was refactored with
translations and now the logout action navs back
to home page. This can be taken much further once a
user account has been setup.